Move style sheet helper to its parent dir
