+++ lib/jsdom/living/css/stylesheets.js
@@ -2,7 +2,7 @@
 const { CSSStyleSheet, StyleSheetList } = require("@cdoublev/css");
 const { legacyHookDecode } = require("@exodus/bytes/encoding.js");
 const whatwgURL = require("whatwg-url");
-const { wrapperForImpl } = require("../../../../generated/idl/utils");
+const { wrapperForImpl } = require("../../../generated/idl/utils");
 
 // TODO: this should really implement https://html.spec.whatwg.org/multipage/links.html#link-type-stylesheet
 // It (and the things it calls) is nowhere close right now.
