+++ lib/jsdom/living/nodes/Document-impl.js
@@ -9,7 +9,7 @@ const { mixin, memoizeQuery } = require("../../utils");
 const { firstChildWithLocalName, firstChildWithLocalNames, firstDescendantWithLocalName } =
   require("../helpers/traversal");
 const whatwgURL = require("whatwg-url");
-const { removeStyleSheet } = require("../css/helpers/stylesheets");
+const { removeStyleSheet } = require("../css/stylesheets");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const eventAccessors = require("../helpers/create-event-accessor");
 const { asciiLowercase, stripAndCollapseASCIIWhitespace } = require("../helpers/strings");
