+++ lib/jsdom/living/nodes/HTMLLinkElement-impl.js
@@ -2,7 +2,7 @@
 const DOMTokenList = require("../../../generated/idl/DOMTokenList");
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const idlUtils = require("../../../generated/idl/utils");
-const { fetchStyleSheet, removeStyleSheet } = require("../css/helpers/stylesheets");
+const { fetchStyleSheet, removeStyleSheet } = require("../css/stylesheets");
 const whatwgURL = require("whatwg-url");
 
 // Important reading: "appropriate times to obtain the resource" in
