+++ lib/jsdom/living/nodes/HTMLStyleElement-impl.js
@@ -1,6 +1,6 @@
 "use strict";
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
-const { addStyleSheet, createStyleSheet, removeStyleSheet, scanForImportRules } = require("../css/helpers/stylesheets");
+const { addStyleSheet, createStyleSheet, removeStyleSheet, scanForImportRules } = require("../css/stylesheets");
 const { childTextContent } = require("../helpers/text");
 const { asciiCaseInsensitiveMatch } = require("../helpers/strings");
 const { fireAnEvent } = require("../helpers/events");
